chore: Centralize the snapshot tests for objects under core To minimize the duplication of snapshot testing code across themes, centralized the object ones in core in a manner similar to how schema tests work with utils and the validators.
