+++ packages/core/testSnap/objectTests.tsx
@@ -1,0 +1,34 @@
+import { ComponentType } from 'react';
+import renderer from 'react-test-renderer';
+import { RJSFSchema } from '@rjsf/utils';
+import validator from '@rjsf/validator-ajv8';
+
+import { FormProps } from '../src';
+
+// export type ObjectRenderCustomOptions = {
+//  checkboxes?: TestRendererOptions;
+// }
+
+export default function arrayTests(Form: ComponentType<FormProps>) {
+  describe('object fields', () => {
+    test('object', () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        properties: {
+          a: { type: 'string', title: 'A' },
+          b: { type: 'number', title: 'B' },
+        },
+      };
+      const tree = renderer.create(<Form schema={schema} validator={validator} />).toJSON();
+      expect(tree).toMatchSnapshot();
+    });
+    test('additionalProperties', () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        additionalProperties: true,
+      };
+      const tree = renderer.create(<Form schema={schema} validator={validator} formData={{ foo: 'foo' }} />).toJSON();
+      expect(tree).toMatchSnapshot();
+    });
+  });
+}
